hack: Add SingleNamespace InstallMode support for the curator-operator
Update the CSV YAML manifest and add support for the SingleNamespace
InstallMode. Currently the koku-metrics-operator doesn't support the
AllNamespaces OperatorGroup InstallMode configuration, which is
problematic in the context of deploying both of those operators in the
same namespace, as we run into scoping problems and OLM cannot roll out
the koku-metrics-operator CSV safely.

The curator-operator has a hard dependency on the koku-metrics-operator,
and that dependency is described in the bundle/metadata/properties.yaml
configuration, which instructs OLM to install koku when that package
is not already present as OLM's resolution model is scoped to individual
namespaces (vs. detecting koku is already installed in a separate
namespace).

Signed-off-by: timflannagan <timflannagan@gmail.com>