Introduce a simple postgres runner
Introduce a simple postgres runner implementation that's responsible for
processing a directory that contains a series of CSV files, parsing
those CSV files for their values, and inserting those CSV values into a
postgres database.

The test-data directory contains a copy of actual koku-metrics-operator
PVC results that were stored in the upload/ directory.

The postgres.yaml file contains the required Kubernetes manifests to
deploy a postgres instance on a vanilla k8s clusters. When interacting
with the postgres runner executable, ensure that the postgres Service
resource has been port-forwarded to a localhost:5432 port or else you'll
get permission denied errors.

Signed-off-by: timflannagan <timflannagan@gmail.com>